intermediate cast to size_t instead of long for convert int->void*
- to fix MSVC warning C4312: 'type cast':
    conversion from 'long' to 'void *' of greater size

+++ src/scip/prop_symmetry.c
@@ -3957,8 +3957,8 @@ SCIP_RETCODE addWeakSBCsSubgroup(
       {
          /* add element from lexorder to hashmap.
           * Use insert, as duplicate entries in lexorder is not permitted. */
-         assert( ! SCIPhashmapExists(varsinlexorder, (void*) (long) (*lexorder)[k]) ); /* Use int as pointer */
-         SCIP_CALL( SCIPhashmapInsertInt(varsinlexorder, (void*) (long) (*lexorder)[k], k) );
+         assert( ! SCIPhashmapExists(varsinlexorder, (void*) (size_t) (*lexorder)[k]) ); /* Use int as pointer */
+         SCIP_CALL( SCIPhashmapInsertInt(varsinlexorder, (void*) (size_t) (*lexorder)[k], k) );
       }
    }
 
@@ -3987,7 +3987,7 @@ SCIP_RETCODE addWeakSBCsSubgroup(
 
       /* If varidx is in lexorder, then it must be the first entry of lexorder. */
       if ( varsinlexorder != NULL
-         && SCIPhashmapExists(varsinlexorder, (void*) (long) varidx)
+         && SCIPhashmapExists(varsinlexorder, (void*) (size_t) varidx)
          && lexorder != NULL && *lexorder != NULL && *maxnvarsorder > 0 && *nvarsorder > 0
          && (*lexorder)[0] != varidx )
          continue;
@@ -4097,13 +4097,13 @@ SCIP_RETCODE addWeakSBCsSubgroup(
             if ( varidx == (*lexorder)[0] )
             {
                /* lexorder is already ok!! */
-               assert( SCIPhashmapExists(varsinlexorder, (void*) (long) varidx) );
+               assert( SCIPhashmapExists(varsinlexorder, (void*) (size_t) varidx) );
             }
             else
             {
                /* Then varidx must not be in the lexorder,
                 * We must add it at the front of the array, and maintain the current order. */
-               assert( ! SCIPhashmapExists(varsinlexorder, (void*) (long) varidx) );
+               assert( ! SCIPhashmapExists(varsinlexorder, (void*) (size_t) varidx) );
 
                ++(*maxnvarsorder);
                ++(*nvarsorder);
